requirements.txt
pydantic
fastapi~=0.65.2
jinja2
uvicorn[standard]
gunicorn
aiofiles
python-dotenv~=0.15.0
psycopg2
fastapi-users[sqlalchemy]
databases[postgresql]~=0.4.3

sqlalchemy~=1.3.24